@@ -16,7 +16,7 @@
     "babel-loader": "6.3.2",
     "babel-preset-es2015": "6.22.0",
     "css-loader": "0.26.1",
-    "eslint": "3.15.0",
+    "eslint": "3.16.0",
     "eslint-loader": "1.6.1",
     "exports-loader": "0.6.3",
     "extract-text-webpack-plugin": "2.0.0-beta.4",
@@ -32,7 +32,7 @@
     "jasmine-core": "2.5.2",
     "jasmine-jquery": "2.1.1",
     "jquery": "3.1.1",
-    "karma": "1.4.1",
+    "karma": "1.5.0",
     "karma-coverage": "1.1.1",
     "karma-firefox-launcher": "1.0.0",
     "karma-jasmine": "1.1.0",
@@ -40,15 +40,15 @@
     "karma-verbose-reporter": "0.0.5",
     "karma-webpack": "2.0.2",
     "node-sass": "4.5.0",
-    "postcss-loader": "1.3.0",
+    "postcss-loader": "1.3.1",
     "raw-loader": "0.5.1",
     "sass-loader": "6.0.1",
     "sinon": "1.17.7",
     "style-loader": "0.13.1",
     "url-loader": "0.5.7",
     "webpack": "2.2.1",
-    "webpack-dev-server": "2.3.0",
-    "webpack-merge": "2.6.1"
+    "webpack-dev-server": "2.4.1",
+    "webpack-merge": "3.0.0"
   },
   "scripts": {
     "build": "gulp build",
